Width-Depth: cleaned up ComplexTime and QubitPool aux classes
