+++ src/Simulation/QCTraceSimulator/QubitMetrics.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Quantum.Simulation.QCTraceSimulatorRuntime
     /// </summary>
     public class QubitTimeMetrics
     {
-        // TODO: Qubit Ids are already available in qubits, but DepthCounter doesn't have access to it
+        // Qubit Ids are already available in qubits, but DepthCounter doesn't have access to it
         // in OnPrimitiveOperation because it's not part of IQCTraceSimulatorListener interface.
         // Consider changing architecture to pass qubits rather than metrics in IQCTraceSimulatorListener.
         public long QubitId { get; }
